full-stack-deepagents/mcp-server-bingchuan/requirements.txt
mcp>=1.26.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.34.0
